Tasked logger: Add per-caller Exception_Occurrence storage.
The last Exception_Occurrence is now stored in the Exception_Occurrences
map on a per-caller basis. The caller is identified by it's Task_Id. The
last Exception_Occurrence can be inquired via Get_Last_Exception.

The last exception occurence is cleared by the next call to Log_Message.
If no new exception occurs it is deleted. Otherwise it is replaced by the
new exception occurrence.