Remove Get_[Facility|Transform] operations.
Operating on facilities/transforms of a logger should be done by
using the respective Update() procedures.

Removing these operations makes handling of facilities and
transforms more robust. It makes it almost impossible for a user
to hold a handle to a facility or transform that has been
deleted.

The only way to hold an invalid handle is to keep the handle of a
facility/transform when inserting it and then calling Delete() on
the logger for that facility/transform.