Active_Logger: Add Shutdown_Helper.
This helper will call Shutdown on the logger given as
discriminant when it goes out of scope.
This relieves the user from having to excplicitly call shutdown
on an instance of Alog active logger when wanting to terminate.

Users must make sure to declare any shutdown helper in a smaller
scope than the active logger on which the helper supposed to
work.