Add controlled map generic.
This generic implements a controlled variant of an indefinite
ordered map.

The memory of an element pointed to by a previously inserted
handle is freed upon calling Delete, Clear or during finalization
of the controlled type. Thus control over object inserted into
this map resides with the controlled map users do not need to
worry about freeing memory.