Tasked_Logger: Add Update entry for facilities.
This entry updates a facility identified by name and calls the
passed process procedure for that facility. Any exception that
might occur during the update is stored and can be retrieved via
a call to Get_Last_Exception().

The passed in procedure is of type 'Facility_Process_Handle'
since only procedure *access types* are allowed as entry
parameters.

To avoid any access check issues the process procedure must be
declared on library level.