Tasked_Logger: Add Iterate entry for facilities.
This entry iterate calls the given facility update handle for all
attached facilities. Any exception that might occur during the
iteration is stored and can be retrieved via a call to
Get_Last_Exception().

As with the update entry the passed in procedure is of type
'Facility_Update_Handle' since only procedure *access types* are
allowed as entry parameters.