Logger: Add per-source loglevel handling.
A source can be specified when logging a message. This allows for
per-source loglevels.

If a source is specified the logger checks if there is an
existing loglevel entry for this source in the sources map. If no
entry is found the message is logged. If an associated loglevel
is found the message is processed only if the specified loglevel
is greater or equal than the one in the map.